% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_linear_functions.R
\name{build_DNAwalk}
\alias{build_DNAwalk}
\title{build_DNAwalk}
\usage{
build_DNAwalk(fname = fname, inpath = inpath, outpath = outpath)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Builds a DNA walk for any fasta file of a DNA sequnece
}
\examples{
function()
}
