% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_linear_functions.R
\name{calc_R2_DNAwalk}
\alias{calc_R2_DNAwalk}
\title{calc_R2_DNAwalk}
\usage{
calc_R2_DNAwalk(
  fname = fname,
  chromosome = chromosome,
  inpath = inpath,
  outpath = outpath
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Calculates the slope and R^2 for any DNAwalk
}
\examples{
function()
}
