% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_linear_functions.R
\name{calc_R2_genes}
\alias{calc_R2_genes}
\title{calc_R2_genes}
\usage{
calc_R2_genes(
  fname = fname,
  chromosome = chromosome,
  inpath = inpath,
  outpath = outpath,
  gff3_path,
  gff3_chromosome
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Calculates an R^2 value for each gene in a gff3 file
}
\examples{
function()
}
