% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_linear_functions.R
\name{calc_R2_TEs}
\alias{calc_R2_TEs}
\title{calc_R2_TEs}
\usage{
calc_R2_TEs(
  fname = fname,
  chromosome = chromosome,
  inpath = inpath,
  outpath = outpath,
  EDTA_path,
  EDTA_chromosome
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Calculates an R^2 value for each TE in a gff3 file from EDTA
}
\examples{
function()
}
