% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{calculate_ranges}
\alias{calculate_ranges}
\title{calculate_ranges}
\usage{
calculate_ranges(
  fname = fname,
  outpath = outpath,
  Shannon_bin_size = 500,
  Shannon_SD = 2,
  RepAbundMax_bin_size = 80,
  RepAbundMax_SD = 2.5,
  RepAbundMin_bin_size = 500,
  RepAbundMin_SD = 2
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Plots the roll sum abundance for all chromosomes. This code can only be run once the rest of the progam (including roll_sum_histogram) is complete.
}
\examples{
function()
}
