% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{cent_finalize}
\alias{cent_finalize}
\title{cent_finalize}
\usage{
cent_finalize(fname = fname, outpath = outpath)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Gathers the centromere predictions from all methods - run at the very end and not automatically
}
\examples{
function()
}
