% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{dna_1_to_wax}
\alias{dna_1_to_wax}
\title{dna_1_to_wax}
\usage{
dna_1_to_wax(dna_1, linerng = NULL)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Makes DNA Walks
}
\examples{
function()
}
