% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{DNAwalks_with_genes}
\alias{DNAwalks_with_genes}
\title{DNAwalks_with_genes}
\usage{
DNAwalks_with_genes(
  chromosome = chromosome,
  fname = fname,
  inpath = inpath,
  outpath = outpath,
  gff3_path = gff3_path,
  cytchr = cytchr,
  start = start,
  end = end
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Creates 2D and 1D  DNA walks of segements of chromosomes showing the genes plotted in blue (forward), red (reverse), black (intergenic) and yellow (overlapping forward and reverse).
}
\examples{
function()
}
