% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{gene_slopes}
\alias{gene_slopes}
\title{gene_slopes}
\usage{
gene_slopes(
  chromosome = chromosome,
  fname = fname,
  inpath = inpath,
  outpath = outpath,
  gff3_path = gff3_path,
  cytchr = cytchr,
  start = start,
  end = end
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Plots graphs of the slopes of each DNA walk for each gene
}
\examples{
function()
}
