% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{imagenan}
\alias{imagenan}
\title{imagenan}
\usage{
imagenan(
  x,
  yline = 3,
  yma = 5,
  topyma = 4,
  xline = 3,
  xma = 6,
  lnumr = 10,
  lnumc = 10,
  lasval = 1,
  cex.axis = 0.7,
  cex.lab = 0.7,
  cex.main = 0.7,
  widths = base::c(5, 1.25),
  heights = base::c(1, 0.5),
  col = grDevices::topo.colors(255),
  outside.below.color = "black",
  outside.above.color = "white",
  na.color = "grey28",
  ...
)
}
\arguments{
\item{x}{2D data to display}

\item{yline}{Number of lines in (left side) to where image begins}

\item{yma}{Number of characters in (left side)}

\item{col}{colour table}

\item{outside.below.color}{Colour below threshold}

\item{outside.above.color}{Colour above threshold}

\item{na.color}{Colour of NAs (see \code{\link[graphics]{par}})}

\item{...}{other plot parameters eg zlim list with min and max range of color table data e.g. zlim=base::c(2,5)}
}
\value{
None
}
\description{
Displays a false colour image of two dimensional data set
}
\examples{
d=base::data.frame(base::cbind(base::c(1:4),base::c(2,5,NA,NA) ))
imagenan(d)
}
