% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{join_chromosome_parts}
\alias{join_chromosome_parts}
\title{join_chromosome_parts}
\usage{
join_chromosome_parts(
  fname = fname,
  chromosome = chromosome,
  inpath = inpath,
  outpath = outpath
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Creates barplots of the spectra and DNAwalks. Needs to be run after the spectra have all been built.
}
\examples{
function()
}
