% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{largeimagesub}
\alias{largeimagesub}
\title{largeimagesub}
\usage{
largeimagesub(
  All_specAve,
  ofi,
  rangebp = NULL,
  rangeseq1 = NULL,
  flaglog = TRUE,
  main = "",
  pngflag = TRUE,
  powval = 1.2,
  inmain = ""
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Plots fourier spectrum before merging
}
\examples{
function()
}
