% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{largeimagesub_NEW}
\alias{largeimagesub_NEW}
\title{largeimagesub_NEW}
\usage{
largeimagesub_NEW(
  All_specAve,
  fname,
  chromosome,
  ofi,
  rangebp = NULL,
  rangeseq1 = NULL,
  flaglog = TRUE,
  main = "",
  pngflag = TRUE,
  powval = 1.2,
  inmain = "",
  repround = TRUE,
  part = FALSE,
  pdf_flag = FALSE
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Plots fourier spectrum before merging
}
\examples{
function()
}
