% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{merge_spectra}
\alias{merge_spectra}
\title{merge_spectra}
\usage{
merge_spectra(
  fname = fname,
  chromosome = chromosome,
  inpath = inpath,
  outpath = outpath
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Merges powers of rounded repeat lengths (currently sums).  Needs to be run after the spectra have all been built and merged for each part.
}
\examples{
function()
}
