% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{PCA_repeats}
\alias{PCA_repeats}
\title{PCA_repeats}
\usage{
PCA_repeats(haplotype_path = haplotype_path, Spp_chromosome = Spp_chromosome)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Plots a PCA with repeat lengths as variables and genome position as individuals, then forms kmeans groups and colours Shannon div plots based on groups
}
\examples{
function()
}
