% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{plot.fourier}
\alias{plot.fourier}
\title{plot.fourier}
\usage{
\method{plot}{fourier}(fourier.series, f.0, ts)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Plots Fourier transform information
}
\examples{
function()
}
