% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{plot.frequency.spectrum}
\alias{plot.frequency.spectrum}
\title{plot.frequency.spectrum}
\usage{
\method{plot}{frequency.spectrum}(fr = NULL, X.k, xlimits = NULL, main = "", plotflag = TRUE)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Plots freq spectrum
}
\examples{
function()
}
