% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{plot.harmonic}
\alias{plot.harmonic}
\title{plot.harmonic}
\usage{
\method{plot}{harmonic}(Xk, i, ts, acq.freq, color = "red")
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Plots harmonics
}
\examples{
function()
}
