% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{plot_all_chromosomes}
\alias{plot_all_chromosomes}
\title{plot_all_chromosomes}
\usage{
plot_all_chromosomes(fname = fname, inpath = inpath, outpath = outpath)
}
\arguments{
\item{fname}{}
}
\value{
plots of all chromosomes
}
\description{
plot Shannon and barplots for all chromosomes
}
\examples{
function()
}
