% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{read_All_list}
\alias{read_All_list}
\title{read_All_list}
\usage{
read_All_list(All_list, maxrow = 1e+05)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Reads All list to make All Spec
}
\examples{
function()
}
