% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{read_walk_list}
\alias{read_walk_list}
\title{read_walk_list}
\usage{
read_walk_list(fullwalklist)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
# Reads list of DNAwalk files
}
\examples{
function()
}
