% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{roll_mean_Shannon}
\alias{roll_mean_Shannon}
\title{roll_mean_Shannon}
\usage{
roll_mean_Shannon(fname = fname, outpath = outpath)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Calculates and plots the Shannon diversity of the Fourier spectra
}
\examples{
function()
}
