% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{roll_sum_histogram}
\alias{roll_sum_histogram}
\title{roll_sum_histogram}
\usage{
roll_sum_histogram(fname = fname, outpath = outpath)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Calculates and plots the roll sum abundance from the summary directory files. This code can only be run once the rest of the progam is complete.
}
\examples{
function()
}
