% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{run_20Mbpimag}
\alias{run_20Mbpimag}
\title{run_20Mbpimag}
\usage{
run_20Mbpimag(
  nam = nam,
  fname = fname,
  inpath = inpath,
  outpath = outpath,
  atflag = TRUE,
  pflag = FALSE,
  plotflag = FALSE,
  writeflag = FALSE
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Creates 20Mbp images of the spectra. Needs to be run after the spectra have all been built.
}
\examples{
function()
}
