% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{run_barplots_bedGraph}
\alias{run_barplots_bedGraph}
\title{run_barplots_bedGraph}
\usage{
run_barplots_bedGraph(
  All_spec,
  nam = nam,
  fname = fname,
  inpath = inpath,
  outpath = outpath,
  chromnum = 0,
  atflag = TRUE,
  numstd = 7,
  numstd2 = 3,
  numrange = 200,
  repeat_range = NULL,
  samplesize = 1,
  fftlength = 5000
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Makes barplots of minimums and maximums for repeats  and bedGraphs of repeat locations
}
\examples{
function()
}
