% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{run_chloroplast}
\alias{run_chloroplast}
\title{run_chloroplast}
\usage{
run_chloroplast(
  nam = nam,
  fname = fname,
  outpath = outpath,
  inpath = inpath,
  startval = NULL,
  endval = NULL,
  fftlength = 5000,
  All_walk = NULL,
  ptset = NULL,
  AT_flag = TRUE,
  sample_every = 1,
  walkflag = FALSE,
  atflag = TRUE,
  waxflag = TRUE,
  printlong = FALSE,
  printdna = FALSE,
  pflag = TRUE,
  plotflag = TRUE,
  writeflag = TRUE
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Creates fourier spectrum of chromosome without plotting
}
\examples{
function()
}
