% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{run_chromosome_loop}
\alias{run_chromosome_loop}
\title{run_chromosome_parallel loop}
\usage{
run_chromosome_loop(
  x,
  nam,
  fname,
  atflag,
  AT_flag,
  startgroup,
  majorlen,
  length_majorgroup,
  name_majorgroup,
  length_submajor,
  name_submajorgroup,
  length_fftgroup,
  name_fftgroup,
  length_minor,
  submajor_nam,
  pflag,
  splitname,
  chr,
  outpathtxt,
  outpathpdf,
  outpathfractal,
  outpathspectra,
  outpathfractal.bed,
  outpathDNAWALK,
  genname1,
  Shortnames,
  shortname,
  in_name,
  dna_1,
  main,
  dn1,
  list_plot,
  full_length,
  start_seq,
  All_list,
  All_listAve,
  fullwalklist,
  fractal_dimlist,
  DNAwalk_FULL,
  plotflag,
  writeflag,
  samplesize
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Loops over 5Mbp in parallel
}
\examples{
function()
}
