% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{run_chromosome_parallel}
\alias{run_chromosome_parallel}
\title{run_chromosome_parallel}
\usage{
run_chromosome_parallel(
  nam = nam,
  fname = fname,
  inpath = inpath,
  outpath = outpath,
  startgroup = 1,
  atflag = TRUE,
  AT_flag = TRUE,
  majorlen = 5e+06,
  length_majorgroup = 1e+06,
  name_majorgroup = "1Mbp",
  length_submajor = 5e+05,
  name_submajorgroup = "500Kbp",
  length_fftgroup = 5000,
  name_fftgroup = "5Kbp",
  length_minor = 25000,
  submajor_nam = "25Kb",
  pflag = FALSE,
  plotflag = TRUE,
  writeflag = TRUE,
  x_cpu = x_cpu,
  samplesize = 50
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Creates fourier spectrum of chromosome with plotting
}
\examples{
function()
}
