% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{run_diversity_plots}
\alias{run_diversity_plots}
\title{run_diversity_plots}
\usage{
run_diversity_plots(
  chromosome = chromosome,
  fname = fname,
  inpath = inpath,
  outpath = outpath
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Creates plots of diversity indices (Shannon, Simpson, Fisher's alpha and Pielou's J).
}
\examples{
function()
}
