% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{run_largeimage}
\alias{run_largeimage}
\title{run_largeimage}
\usage{
run_largeimage(
  All_spec,
  nam = nam,
  fname = fname,
  inpath = inpath,
  outpath = outpath,
  atflag = TRUE,
  rangeseq = NULL,
  pngflag = TRUE,
  samplesize = 1,
  r0 = base::c(35, 2000),
  r1 = base::c(15, 35),
  r2 = base::c(2, 8),
  inmain = ""
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Reads All Spec into the full chromosome fourier spectrum
}
\examples{
function()
}
