% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{run_long_repeats_NEW}
\alias{run_long_repeats_NEW}
\title{run_long_repeats_NEW}
\usage{
run_long_repeats_NEW(
  nam = nam,
  fname = fname,
  inpath = inpath,
  outpath = outpath,
  AT_flag = TRUE,
  atflag = TRUE,
  pflag = FALSE,
  plotflag = FALSE,
  writeflag = FALSE
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
# Runs large fft windows to look for longer repeats
}
\examples{
function()
}
