% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{run_one_repeat}
\alias{run_one_repeat}
\title{run_one_repeat}
\usage{
run_one_repeat(
  repeat_val,
  All_spec,
  chromnam,
  numstd = 5,
  colrange = NULL,
  sortstr = FALSE
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Finds peaks in one repeat length
}
\examples{
function()
}
