% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{run_plot_chromosome}
\alias{run_plot_chromosome}
\title{run_plot_chromosome}
\usage{
run_plot_chromosome(
  nam = nam,
  fname = fname,
  inpath = inpath,
  outpath = outpath,
  atflag = TRUE,
  pflag = FALSE,
  plotflag = FALSE,
  writeflag = FALSE
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Creates fourier spectrum of chromosome with plotting
}
\examples{
function()
}
