% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{run_plot_chromosome_parallel}
\alias{run_plot_chromosome_parallel}
\title{run_plot_chromosome_parallel}
\usage{
run_plot_chromosome_parallel(
  nam = nam,
  fname = fname,
  inpath = inpath,
  AT_flag = TRUE,
  outpath = outpath,
  atflag = TRUE,
  pflag = FALSE,
  plotflag = FALSE,
  writeflag = FALSE,
  x_cpu = x_cpu
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Creates fourier spectrum of chromosome
}
\examples{
function()
}
