% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{run_sum_bp}
\alias{run_sum_bp}
\title{run_sum_bp}
\usage{
run_sum_bp(
  repeat_val,
  All_spec,
  chromnam,
  numstd2,
  numrange = 1500,
  pflag = FALSE
)
}
\arguments{
\item{input}{dataset}
}
\value{
output dataset
}
\description{
Finds peaks in one repeat length
}
\examples{
function()
}
