% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{run_summary_hist}
\alias{run_summary_hist}
\title{run_summary_hist}
\usage{
run_summary_hist(
  chromosome = chromosome,
  fname = fname,
  inpath = inpath,
  outpath = outpath,
  atflag = TRUE
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Creates barplots of the spectra and DNAwalks. Needs to be run after the spectra have all been built.
}
\examples{
function()
}
