% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{run_summary_plots}
\alias{run_summary_plots}
\title{run_summary_plots}
\usage{
run_summary_plots(
  chromosome = chromosome,
  fname = fname,
  inpath = inpath,
  outpath = outpath
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Creates spectra and DNAwalks plots for the whole chromosome. Needs to be run after the spectra and DNAwalk have all been built.
}
\examples{
function()
}
