% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{run_summary_plots_range}
\alias{run_summary_plots_range}
\title{run_summary_plots_range}
\usage{
run_summary_plots_range(
  fname = fname,
  chromosome = chromosome,
  inpath = inpath,
  outpath = outpath,
  startbp = startbp,
  endbp = endbp
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Creates spectra and DNAwalks plots for a specific range. Needs to be run after the spectra and DNAwalk have all been built.
}
\examples{
function()
}
