% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{runs_run_barplots}
\alias{runs_run_barplots}
\title{runs_run_barplots}
\usage{
runs_run_barplots(
  nam = nam,
  fname = fname,
  inpath = inpath,
  outpath = outpath,
  atflag = TRUE,
  pflag = FALSE,
  plotflag = FALSE,
  writeflag = FALSE
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Creates barplots of the spectra and DNAwalks. Needs to be run after the spectra have all been built.
}
\examples{
function()
}
