% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{walk_and_plot}
\alias{walk_and_plot}
\title{walk_and_plot}
\usage{
walk_and_plot(
  tdna,
  main = "",
  ptset = NULL,
  rng = NULL,
  seqlen = NULL,
  sample_every = 1,
  fullname = NULL,
  All_walk = NULL,
  walkflag = TRUE,
  fracflag = FALSE,
  AT_flag = TRUE,
  atflag = TRUE,
  spar = 0.6,
  startval = 1,
  endval = NULL,
  printlong = FALSE,
  waxflag = FALSE,
  freqerrorflag = TRUE,
  pflag = TRUE,
  plotflag = TRUE,
  writeflag = TRUE
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Plots DNA Walks
}
\examples{
function()
}
