% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnawalk_fft_functions.R
\name{write_All_spec_DNAwalk}
\alias{write_All_spec_DNAwalk}
\title{write_All_spec_DNAwalk}
\usage{
write_All_spec_DNAwalk(
  nam = nam,
  fname = fname,
  chromosome = chromosome,
  inpath = inpath,
  outpath = outpath,
  pflag = FALSE,
  plotflag = FALSE,
  writeflag = FALSE
)
}
\arguments{
\item{nam}{input dataset}
}
\value{
output dataset
}
\description{
Creates barplots of the spectra and DNAwalks. Needs to be run after the spectra have all been built.
}
\examples{
function()
}
